Separated emoji reactions into ravestate_emoji module.
